+++ include/vm/table.h
@@ -23,6 +23,6 @@ int table_set(Table* table, VmString* key, Value value);
 void table_add_all(Table* from, Table* to);
 int table_get(Table* table, VmString* key, Value* value);
 int table_delete(Table* table, VmString* key);
-VmString* table_find_string(Table* table, const char* chars, int length, Hash hash);
+VmString* table_find_string(Table* table, const char* chars, size_t length, Hash hash);
 
 #endif
